finsihish fuction

+++ IIIDShapes/Sphere.java
@@ -8,12 +8,12 @@ public class Sphere extends IIIDShape{
 
     @Override
     public double Volume() {
-        return Math.round(((4/3) * Math.PI * Math.pow(radius, 3)) * 100) / 100;
+        return (double) Math.round(((4/3) * Math.PI * Math.pow(radius, 3)) * 100) / 100;
     }
 
     @Override
     public double SA() {
-        return Math.round((4 * (Math.PI) * Math.pow(radius, 2)) * 100) / 100;
+        return  (double) Math.round((4 * (Math.PI) * Math.pow(radius, 2)) * 100) / 100;
     }
 
     @Override
